requirements.txt
pytest==6.2.5
pytest-cov==2.12.1
requests==2.22.0
selenium==3.141.0
pytest-reporter
pytest-html
flake8
webdriver-manager
allure-pytest
